Projekt 1/client/app/requirements.txt
flask
uuid
pyjwt
redis
requests
flask-swagger-ui
python-dotenv
python-jose
flask-cors
six